move more s3 specific code into storage/s3.rs
reduce cfg flags in storage/mod.rs
